refactor(messaging): move quoted file helpers into messaging
